max-agent-v2/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Database
supabase==2.9.0
sqlalchemy==2.0.35

# AI Services
openai==1.54.0
tiktoken==0.8.0
openai-whisper==20240930
setuptools-rust==1.10.2
langchain-core>=1.2.11

# WhatsApp
requests==2.32.3
aiohttp==3.10.10

# Admin Panel
flask==3.0.3
flask-cors==5.0.0
jinja2==3.1.4

# Notifications
python-telegram-bot==21.7
resend==2.4.0

# Task Queue
celery==5.4.0
redis==5.2.0

# Utils
python-dateutil==2.9.0
pytz==2024.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Validation
dnspython==2.6.1
